Fix bug with data type in l1 reg loss

+++ nichecompass/modules/losses.py
@@ -356,7 +356,7 @@ def compute_masked_l1_reg_loss(model: nn.Module,
         L1 regularization loss for the masked decoder layer weights.
     """
     if only_target_genes & (model.node_label_method_ != "self"):
-        param_end_gene_idx = (model.n_output_genes_ / 2)
+        param_end_gene_idx = int(model.n_output_genes_ / 2)
     else:
         param_end_gene_idx = None
